FL Model/Codes/requirements.txt
flwr>=1.4.0
numpy
pandas
scikit-learn
matplotlib
tensorflow-lite